+++ src/com/computerDatabase/DAO/ComputerDao.java
@@ -7,9 +7,9 @@ import com.computerDatabase.model.*;
 public interface ComputerDao {
 	
     void create( Computer computer ) throws DAOException;
-    void delete(int id) throws DAOException;
+    void delete(long id) throws DAOException;
     void update( Computer computer ) throws DAOException;
     Collection<Computer> list() throws DAOException;
-    Computer details( int id ) throws DAOException;
+    Computer details( long id ) throws DAOException;
 	
 }
